third_party/blink/web_tests/harness-tests/console_logging-expected.txt

CONSOLE MESSAGE: Single line message
CONSOLE MESSAGE: Multi-line
message
This is a testharness.js-based test.
[FAIL] Test needs to fail so that an expected.txt file is written out and be examined for console messages.
  assert_true: expected true got false
Harness: the test ran to completion.
